oxish: move Session::from_message() to unix::resume()
